Moved definition to its own file
